% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOVenn.R
\name{circleFun}
\alias{circleFun}
\title{circleFun}
\usage{
circleFun(center = c(0, 0), diameter = 1, npoints = 100)
}
\arguments{
\item{center}{Center point of the circle}

\item{diameter}{Defines the diameter of the circle}

\item{npoints}{Number of points used to draw the circle}
}
\description{
circleFun
}
