% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOclass.R
\docType{class}
\name{gclus-class}
\alias{gclus-class}
\alias{gclus}
\title{An S4 class to represent gene cluster}
\description{
Store cluster sets in a tbl object.
}
\section{Slots}{

\describe{
\item{\code{tbl}}{tbl_df object}
}}

