% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOmethods.R
\docType{methods}
\name{GO,gclus,gset,logical,logical-method}
\alias{GO,gclus,gset,logical,logical-method}
\title{calculate OR
((g present in clus c and pathway p)/(g present in clus c but absent in pathway p))((g absent in clus c but present in pathway p)/(g absent in clus c and absent in pathway p))
reshuffle gene sets 100,000 times, and calcualte OR
calculate Pval
calculate Padj}
\usage{
\S4method{GO}{gclus,gset,logical,logical}(gclus.obj, gset.obj,
  filterPADJ = TRUE, filterOR = TRUE)
}
\description{
calculate OR
((g present in clus c and pathway p)/(g present in clus c but absent in pathway p))((g absent in clus c but present in pathway p)/(g absent in clus c and absent in pathway p))
reshuffle gene sets 100,000 times, and calcualte OR
calculate Pval
calculate Padj
}
