% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOgeneric.R, R/GOmethods.R
\docType{methods}
\name{GO}
\alias{GO}
\alias{GO,gclus,gset-method}
\title{Odds Ratio Analysis}
\usage{
GO(gclus.obj, gset.obj, filterPADJ = TRUE, filterOR = TRUE,
  Padj.cutoff = 0.05)

\S4method{GO}{gclus,gset}(gclus.obj, gset.obj, filterPADJ = TRUE,
  filterOR = TRUE, Padj.cutoff = 0.05)
}
\arguments{
\item{gclus.obj}{A \code{gclus} object.}

\item{gset.obj}{A \code{gset} object.}

\item{filterPADJ}{Logical, whether to filter with a P value threshold set by \code{Padj.cutoff}.}

\item{filterOR}{Logical, whether to filter out infinite ORs.}

\item{Padj.cutoff}{Numeric, specififying significance cutoff on adjust P value.}
}
\value{
A list with two entries \code{go_res.obj} and \code{go_set.obj}.
}
\description{
Odds ratio and P value are computed from \code{\link{fisher.test}}.
}
\details{
See \code{\link{fisher.test}}. An alternative and slower approach is to compute odds ratio by a contingency table and estiamte P value from repeated reshuffle analysis.
}
\examples{
\dontrun{
library(GOplot)
gset.obj <- selDB(major="C2.CP", minor="Reactome", type="symbols", species="mouse")
gclus.obj <- new("gclus", tbl=tibble:::as_tibble(data.frame( gene = c("Nanog","Rpl3","Rpl4","Mbl2","Ubr1","Herc2","Asb4","Rnf123","Klf4","Uba5"), clus = factor(c(rep(1,6),rep(2,4))) )) )
res.list <- GO(gclus.obj, gset.obj, filterPADJ=FALSE, filterOR=TRUE)
go_set.obj <- res.list$go_set.obj
go_res.obj <- res.list$go_res.obj
write_GO(go_set.obj, go_res.obj, 'output_txt')
simi(go_set.obj, go_res.obj, 'output_fig')
}
}
