% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_output_to_df.R
\name{go_output_to_df}
\alias{go_output_to_df}
\title{A munging function to turn Gene Ontology outputs into data frame}
\usage{
go_output_to_df(go.out)
}
\arguments{
\item{go.out}{the output from running enrichGO() or run_gene_ontology()}
}
\description{
A munging function to turn Gene Ontology outputs into data frame
}
\examples{
go_output_to_df()
}
