% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOclass.R
\docType{class}
\name{go_res-class}
\alias{go_res-class}
\alias{go_res}
\title{An S4 class to represent GO enrichment analysis result}
\description{
Store GO results in a tbl object.
}
\section{Slots}{

\describe{
\item{\code{tbl}}{tbl_df object}
}}

