% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOclass.R
\docType{class}
\name{gset-class}
\alias{gset-class}
\alias{gset}
\title{An S4 class to represent gene sets.}
\description{
Store gene sets in a tbl object.
}
\section{Slots}{

\describe{
\item{\code{tbl}}{tbl_df object  contains = c("tbl_df"),}
}}

