% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_go_barchart.R
\name{plot_go_barchart}
\alias{plot_go_barchart}
\title{Plot gene ontology outputs}
\usage{
plot_go_barchart(go.df, n = 30)
}
\arguments{
\item{go.df}{a data frame with at least Ontology, Description, and p.adjust}
}
\description{
Plot gene ontology outputs
}
\details{
Cleaner looking version
}
