% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gene_ontology.R
\name{run_gene_ontology}
\alias{run_gene_ontology}
\title{A wrapper around clusterProfiler::enrichGO}
\usage{
run_gene_ontology(genes)
}
\arguments{
\item{genes}{a vector of ENSEMBL (the ones that start ENS000000...) strings}
}
\description{
This function
}
