% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOconst.R
\name{selDB}
\alias{selDB}
\title{Select db for GO}
\usage{
selDB(major = c("C2.CP", "C3.MIR", "C3.TFT", "C5.BP", "C5.CC"),
  minor = c("Reactome"), type = c("symbols", "entrez"),
  species = c("human", "mouse"))
}
\arguments{
\item{major, }{the major category.}

\item{minor, }{a character string to subset through grep. Set NA to ignore.}

\item{type, }{either symbols or entrez.}

\item{species, }{either human or mouse.}
}
\value{
A \code{gset} object.
}
\description{
Initate gset class by making use of gene sets from package \code{msgidb}
}
\details{
For details about the gene sets, please see package \code{msgidb}.
}
