% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOmethods.R
\docType{methods}
\name{simi,gset,go_res,character-method}
\alias{simi,gset,go_res,character-method}
\title{estimate similarity between gene set terms}
\usage{
\S4method{simi}{gset,go_res,character}(go_set.obj, go_res.obj, nms)
}
\description{
estimate similarity between gene set terms
}
