% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOgeneric.R, R/GOmethods.R
\docType{methods}
\name{simi}
\alias{simi}
\alias{simi,gset,go_res,character-method}
\title{Presentation of similar gene sets}
\usage{
simi(go_set.obj, go_res.obj, nms)

\S4method{simi}{gset,go_res,character}(go_set.obj, go_res.obj, nms)
}
\arguments{
\item{go_set.obj}{A go_set object.}

\item{go_res.obj}{A go_res object.}

\item{nms}{A directory with prefix for output.}
}
\description{
Estimate similarity between gene set terms and plot clustering
}
