% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOmethods.R
\docType{methods}
\name{write_GO,gset,go_res,character-method}
\alias{write_GO,gset,go_res,character-method}
\title{output GO results}
\usage{
\S4method{write_GO}{gset,go_res,character}(go_set.obj, go_res.obj, nms)
}
\description{
output GO results
}
