% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOgeneric.R, R/GOmethods.R
\docType{methods}
\name{write_GO}
\alias{write_GO}
\alias{write_GO,gset,go_res,character-method}
\title{Output GO Results}
\usage{
write_GO(go_set.obj, go_res.obj, nms)

\S4method{write_GO}{gset,go_res,character}(go_set.obj, go_res.obj, nms)
}
\arguments{
\item{go_set.obj}{A go_set object.}

\item{go_res.obj}{A go_res object.}

\item{nms}{A directory with prefix for output.}
}
\description{
Write GO results into a tsv file
}
