% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GOCore.R
\name{draw_table}
\alias{draw_table}
\title{draw table}
\usage{
draw_table(data, col = "")
}
\arguments{
\item{data}{data frame; GO ID and GO term}

\item{col}{character; defines color of table header}
}
\description{
draw table
}

