% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOclasss.R
\docType{class}
\name{gset-class}
\alias{gset-class}
\alias{gset}
\title{import GOtest
import tidyverse
import ggplot2
import ggdendro
import RcolorBrewer
import multiplot
genes-clusters 1-to-1 relation
gene sets: list (gene sets) of vectors (genes)
per cluster GO terms: GOlabel, OR, pvalue, qvalue
per cluster for a subset of GO terms: GO relationship matrix
export gset}
\description{
import GOtest
import tidyverse
import ggplot2
import ggdendro
import RcolorBrewer
import multiplot
genes-clusters 1-to-1 relation
gene sets: list (gene sets) of vectors (genes)
per cluster GO terms: GOlabel, OR, pvalue, qvalue
per cluster for a subset of GO terms: GO relationship matrix
export gset
}
