% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOconst.R
\name{selDB}
\alias{selDB}
\title{select db for GO
initate gset class
minor must be set to NULL to de-select minor DB.}
\usage{
selDB(major = c("C2.CP", "C3.MIR", "C3.TFT", "C5.BP", "C5.CC"),
  minor = c("Reactome"), type = c("symbols", "entrez"),
  species = c("human", "mouse"))
}
\description{
select db for GO
initate gset class
minor must be set to NULL to de-select minor DB.
}
